Delete method for Train, Staton, Passenger, Ticket, Route

+++ bookingtrain/src/main/java/com/example/bookingtrain/repository/StationRepository.java
@@ -3,6 +3,6 @@ package com.example.bookingtrain.repository;
 import com.example.bookingtrain.model.Station;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface StationRepository extends JpaRepository<Station, Long> {
+public interface StationRepository extends JpaRepository<Station, Integer> {
 
 }
