+++ bookingtrain/src/main/java/com/example/bookingtrain/repository/TrainRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TrainRepository extends JpaRepository<Train, Long> {
+public interface TrainRepository extends JpaRepository<Train, Integer> {
 
     Train findByTrainName(String trainName);
     Train findByTrainCode(String trainCode);
